Add `arrays` and `store` instead of `yield`

+++ language/src/main/java/net/wrap_trap/truffle_arrow/language/truffle/node/ExprPlus.java
@@ -7,7 +7,7 @@ import com.oracle.truffle.api.nodes.NodeInfo;
 public abstract class ExprPlus extends ExprBinary {
 
   @Specialization
-  protected long add(int left, int right) {
+  protected int add(int left, int right) {
     return left + right;
   }
